+++ src/components/custom/misc/index.js
@@ -1,2 +1,2 @@
 export { AudioClipSamples } from "./audio-clip-samples";
-export { L13ASummersDay } from "./daily-routine-poem";
+export { L13ASummersDay as DailyRoutineASummersDay } from "./daily-routine-poem";