+++ src/components/custom/pronunciation/index.js
@@ -1,32 +1,31 @@
 export {
-	LO1Demystify1,
-	LO1Demystify2,
-	LO1Demystify3,
-	LO1Demystify4,
+	LO1Demystify1 as FirstContactPronunciationNasalOnOm,
+	LO1Demystify2 as FirstContactPronunciationFrenchR,
+	LO1Demystify3 as FirstContactPronunciationOiSound,
+	LO1Demystify4 as FirstContactPronunciationSilentH,
 } from "./first-contact-pronunciation";
 
 export {
-	LO2Pronunciation1,
-	LO2Pronunciation2,
+	LO2Pronunciation1 as AboutMePronunciationSilentE,
+	LO2Pronunciation2 as AboutMePronunciationFrenchTh,
 } from "./about-me-pronunciation";
 
 export {
-	LO3Demystify,
-	LO3Pronunciation1,
-	LO3Pronunciation2,
-	LO3Pronunciation3,
+	LO3Pronunciation1 as OriginsAndLanguagesPronunciationSilentEnt,
+	LO3Pronunciation2 as OriginsAndLanguagesPronunciationFinalConsonantsLiaison,
+	LO3Pronunciation3 as OriginsAndLanguagesPronunciationEhSound,
 } from "./origins-and-languages-pronunciation";
 
 export {
-	LO4Pronunciation1,
-	LO4Pronunciation2,
-	LO4Pronunciation3,
+	LO4Pronunciation1 as CurrentLocationPronunciationNasalInAin,
+	LO4Pronunciation2 as CurrentLocationPronunciationRelatedSpellings,
+	LO4Pronunciation3 as CurrentLocationPronunciationFinalEnEns,
 } from "./current-location-pronunciation";
 
-export { LO5Demystify } from "./house-and-home-pronunciation";
-export { LO6Demystify } from "./family-friends-and-neighbours-pronunciation";
-export { LO7Demystify } from "./opinions-matter-pronunciation";
-export { LO8Demystify } from "./free-time-pronunciation";
-export { LO9Demystify } from "./phoning-in-france-pronunciation";
-export { L10Demystify } from "./making-arrangements-pronunciation";
-export { L12Demystify } from "./shopping-in-the-market-pronunciation";
+export { LO5Demystify as HouseAndHomePronunciation } from "./house-and-home-pronunciation";
+export { LO6Demystify as FamilyFriendsAndNeighboursPronunciation } from "./family-friends-and-neighbours-pronunciation";
+export { LO7Demystify as OpinionsMatterPronunciation } from "./opinions-matter-pronunciation";
+export { LO8Demystify as FreeTimePronunciation } from "./free-time-pronunciation";
+export { LO9Demystify as PhoningInFrancePronunciation } from "./phoning-in-france-pronunciation";
+export { L10Demystify as MakingArrangementsPronunciation } from "./making-arrangements-pronunciation";
+export { L12Demystify as ShoppingInTheMarketPronunciation } from "./shopping-in-the-market-pronunciation";